Add parkingLotId to ParkingLotInfoDto and ParkingLotEntity, create SubscribeModel and PageBottomPopupViewModel
This commit adds a new field called parkingLotId to the ParkingLotInfoDto and ParkingLotEntity classes. It also creates a new SubscribeModel class that handles subscribing to notifications for a specific parking lot. Finally, it creates a new PageBottomPopupViewModel class that uses the SubscribeModel to handle subscribing from the bottom popup view. A subscription button is added to the custom_view_pager.xml layout file as well.